Fix int pos = std::string::find() et al problems
git-svn-id: https://svn.isis.vanderbilt.edu/GMESRC/trunk@1677 b35830b6-564f-0410-ac5f-b387e50cb686

+++ GME/ConstraintManager/GMEConstraintEx.cpp
@@ -71,7 +71,7 @@ namespace OclGme
 		}
 
 		std::string strDName = GetDefinedName();
-		int iPos = strDName.rfind( "::" );
+		size_t iPos = strDName.rfind( "::" );
 
 		if ( iPos == std::string::npos ) {
 			strContext = "?";
